+++ json.rb
@@ -8,7 +8,7 @@ def create_files
   Dir.mkdir('./storage')
   FileUtils.touch('./storage/book.json')
   FileUtils.touch('./storage/label.json')
-  FileUtils.touch('./storage/music_albums.json')
+  FileUtils.touch('./storage/music.json')
   FileUtils.touch('./storage/genre.json')
   FileUtils.touch('./storage/game.json')
   FileUtils.touch('./storage/author.json')
