+++ storage/book.json
@@ -1,0 +1,1 @@
+[{"ref":"#<Book:0x00000001150865b8>","value":{"id":11,"genre":null,"author":null,"label":null,"publish_date":null,"archived":false,"publisher":"hyads","cover_state":"good"}}]
