+++ storage/game.json
@@ -1,0 +1,1 @@
+[{"ref":"#<Game:0x0000000115085668>","value":{"id":120,"genre":null,"author":null,"label":null,"publish_date":"2022/02/02","archived":false,"multiplayer":"y","last_played_at":"2022/02/02"}}]
