+++ storage/music.json
@@ -1,0 +1,1 @@
+[{"ref":"#<MusicAlbum:0x000000011508fd20>","value":{"id":83,"genre":null,"author":null,"label":null,"publish_date":"2023/03/03","archived":false,"on_spotify":false}}]
